requirements.txt
ansitable==0.11.4
appnope==0.1.4
asttokens==3.0.0
casadi==3.6.7
cfgv==3.4.0
colored==2.3.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
distlib==0.3.9
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
identify==2.6.9
ipykernel==6.29.5
ipython==8.34.0
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
-e git+ssh://git@github.com/micahreich/koopman.git@1ae7bdba6389f214d27196f9bea5362c940c3d62#egg=koopman
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
ml_dtypes==0.5.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pre_commit==4.2.0
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.3.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
spatialmath-python==1.1.14
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
virtualenv==20.29.3
wcwidth==0.2.13
